ISS61 - refactor police report module
