+++ python_src/models/police_report/__init__.py
@@ -1,0 +1,5 @@
+from .report import Report, CAT_A_DATA, CAT_B_DATA, CAT_C_DATA, CAT_D_DATA, categorize_report
+from .reportA import ReportTypeA
+from .reportB import ReportTypeB
+from .reportC import ReportTypeC
+from .reportD import ReportTypeD
