+++ python_src/tests/runner.py
@@ -4,7 +4,6 @@ from test_script_factory import ScriptVisualizationTest
 from test_road_profiler import RoadProfilerTest
 from test_simulation_score import TestSimScore
 
-
 if __name__ == "__main__":
     runner = unittest.TextTestRunner(verbosity=2)
     loader = unittest.TestLoader()
